backend/iprs-mock/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
Markdown>=3.4
django-filter>=23.2
Pillow>=10.0.0
numpy>=1.24.0
